+++ app/package.json
@@ -31,14 +31,15 @@
     "@radix-ui/react-select": "^2.1.2",
     "@radix-ui/react-separator": "^1.1.0",
     "@radix-ui/react-slot": "^1.1.0",
+    "@reown/appkit": "^1.5.2",
+    "@reown/appkit-adapter-wagmi": "^1.5.2",
     "@sentry/nextjs": "^8.35.0",
     "@sentry/profiling-node": "^8.35.0",
     "@snowbridge/api": "^0.1.23",
     "@snowbridge/contract-types": "^0.1.23",
     "@talismn/connect-components": "^1.1.8",
-    "@tanstack/react-query": "^5.59.15",
+    "@tanstack/react-query": "^5.61.5",
     "@vercel/analytics": "^1.3.1",
-    "@web3modal/wagmi": "^5.1.11",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cypress-wait-until": "^3.0.2",
