+++ app/src/app/layout.tsx
@@ -1,15 +1,12 @@
 import Navbar from '@/components/NavBar'
 import NotificationSystem from '@/components/NotificationSystem'
 import { dazzed } from '@/components/fonts/fonts'
-import { config } from '@/config'
-import Web3ModalProvider from '@/context'
+import ReownProvider from '@/context'
 import { TURTLE_CONFIG } from '@/utils/turle.config'
 import { NextUIProvider } from '@nextui-org/react'
 import { Analytics } from '@vercel/analytics/react'
 import type { Metadata } from 'next'
-import { headers } from 'next/headers'
 import { twMerge } from 'tailwind-merge'
-import { cookieToInitialState } from 'wagmi'
 import './globals.css'
 
 export const metadata: Metadata = {
@@ -27,16 +24,14 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode
 }>) {
-  const initialState = cookieToInitialState(config, headers().get('cookie'))
-
   return (
     <html lang="en" className="h-full">
       <body className={twMerge(dazzed.variable, 'min-h-full bg-turtle-tertiary font-dazzed')}>
         <Navbar />
         <NotificationSystem />
-        <Web3ModalProvider initialState={initialState}>
+        <ReownProvider>
           <NextUIProvider>{children}</NextUIProvider>
-        </Web3ModalProvider>
+        </ReownProvider>
         <Analytics />
       </body>
     </html>
