+++ app/src/components/Transfer.tsx
@@ -21,13 +21,11 @@ import Button from './Button'
 import ChainSelect from './ChainSelect'
 import Credits from './Credits'
 import SendButton from './SendButton'
-import SubstrateWalletModal from './SubstrateWalletModal'
 import { AlertIcon } from './svg/AlertIcon'
 import { SwapChains } from './SwapFromToChains'
 import Switch from './Switch'
 import TokenAmountSelect from './TokenAmountSelect'
 import TxSummary from './TxSummary'
-import WalletButton from './WalletButton'
 
 const Transfer: FC = () => {
   const { snowbridgeContext } = useSnowbridgeContext()
@@ -151,7 +149,7 @@ const Transfer: FC = () => {
               options={getAllowedSourceChains(environment)}
               floatingLabel="From"
               placeholder="Source"
-              trailing={<WalletButton walletType={sourceChain?.walletType} />}
+              /* trailing={<WalletButton walletType={sourceChain?.walletType} />} */
               walletAddress={sourceWallet?.sender?.address}
               className="z-50"
               disabled={transferStatus !== 'Idle'}
@@ -214,14 +212,14 @@ const Transfer: FC = () => {
               manualRecipient={manualRecipient}
               onChangeManualRecipient={handleManualRecipientChange}
               error={manualRecipient.enabled ? manualRecipientError : ''}
-              trailing={
+              /* trailing={
                 // TODO: support all address types
                 !manualRecipient.enabled &&
                 sourceChain?.supportedAddressTypes.at(0) !==
                   destinationChain?.supportedAddressTypes.at(0) && (
                   <WalletButton walletType={destinationChain?.walletType} />
                 )
-              }
+              } */
               walletAddress={destinationWallet?.sender?.address}
               className="z-30"
               disabled={transferStatus !== 'Idle' || !sourceChain || !tokenAmount?.token}
@@ -351,7 +349,6 @@ const Transfer: FC = () => {
       />
 
       <Credits />
-      <SubstrateWalletModal />
     </form>
   )
 }
