+++ app/src/context/index.tsx
@@ -1,18 +1,46 @@
 'use client'
-import { config } from '@/config'
-import { projectId } from '@/utils/env'
+import { Environment } from '@/store/environmentStore'
+import { isDevelopment, projectId, vercelDomain } from '@/utils/env'
+import { createAppKit } from '@reown/appkit'
+import { WagmiAdapter } from '@reown/appkit-adapter-wagmi'
+import { mainnet, sepolia } from '@reown/appkit/networks'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { createWeb3Modal } from '@web3modal/wagmi/react'
 import { ReactNode } from 'react'
-import { State, WagmiProvider } from 'wagmi'
-import { colors } from '../../tailwind.config'
+import { WagmiConfig } from 'wagmi'
 
 // Setup queryClient
-const queryClient = new QueryClient()
+export const queryClient = new QueryClient()
 if (!projectId) throw new Error('Project ID is not defined')
 
+// Get projectId at https://cloud.walletconnect.com
+if (!projectId) throw new Error('Project ID is not defined')
+
+const vercelUrl = vercelDomain ? `https://${vercelDomain}` : ''
+export const DWELLIR_KEY = process.env.NEXT_PUBLIC_DWELLIR_KEY
+
+const metadata = {
+  name: 'turtle-app',
+  description: 'Token transfers done right',
+  url: isDevelopment ? 'http://localhost:3000' : vercelUrl, // domain must be allowed in WalletConnect Cloud
+  icons: ['https://avatars.githubusercontent.com/u/179229932'],
+}
+
+export const wagmiAdapter = new WagmiAdapter({
+  networks: process.env.NEXT_PUBLIC_ENVIRONMENT === Environment.Testnet ? [sepolia] : [mainnet],
+  projectId,
+})
+
 // Create modal
-export const wallet = createWeb3Modal({
+createAppKit({
+  adapters: [wagmiAdapter],
+  networks: process.env.NEXT_PUBLIC_ENVIRONMENT === Environment.Testnet ? [sepolia] : [mainnet],
+  metadata: metadata,
+  projectId,
+  features: {
+    analytics: true,
+  },
+})
+/* export const wallet = createWeb3Modal({
   wagmiConfig: config,
   projectId,
   enableAnalytics: true,
@@ -21,18 +49,12 @@ export const wallet = createWeb3Modal({
   themeVariables: {
     '--w3m-accent': colors['turtle-primary'],
   },
-})
+}) */
 
-export default function Web3ModalProvider({
-  children,
-  initialState,
-}: {
-  children: ReactNode
-  initialState?: State
-}) {
+export default function ReownProvider({ children }: { children: ReactNode }) {
   return (
-    <WagmiProvider config={config} initialState={initialState}>
+    <WagmiConfig config={wagmiAdapter.wagmiConfig}>
       <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-    </WagmiProvider>
+    </WagmiConfig>
   )
 }
